+++ core/src/main/java/com/moneydance/modules/features/paypalimporter/CoreModule.java
@@ -16,6 +16,7 @@ import dagger.Provides;
         com.moneydance.modules.features.paypalimporter.bootstrap.FactoryModule.class,
         com.moneydance.modules.features.paypalimporter.controller.FactoryModule.class,
         com.moneydance.modules.features.paypalimporter.domain.FactoryModule.class,
+        com.moneydance.modules.features.paypalimporter.model.FactoryModule.class,
         com.moneydance.modules.features.paypalimporter.service.FactoryModule.class,
         com.moneydance.modules.features.paypalimporter.util.FactoryModule.class,
 })