+++ support/src/main/java/com/moneydance/modules/features/paypalimporter/SupportComponent.java
@@ -6,6 +6,8 @@ package com.moneydance.modules.features.paypalimporter;
 import com.moneydance.apps.md.controller.FeatureModuleContext;
 import com.moneydance.modules.features.paypalimporter.controller.ViewController;
 import com.moneydance.modules.features.paypalimporter.domain.DateConverter;
+import com.moneydance.modules.features.paypalimporter.model.AccountFilter;
+import com.moneydance.modules.features.paypalimporter.model.IAccountBook;
 import com.moneydance.modules.features.paypalimporter.service.ServiceProvider;
 import com.moneydance.modules.features.paypalimporter.util.Localizable;
 import com.moneydance.modules.features.paypalimporter.util.Preferences;
@@ -37,4 +39,8 @@ public interface SupportComponent {
     DateConverter dateConverter();
 
     ServiceProvider serviceProvider();
+
+    IAccountBook accountBook();
+
+    AccountFilter accountFilter();
 }