+++ support/src/test/java/com/moneydance/modules/features/paypalimporter/controller/TransactionSearchRequestHandlerTest.java
@@ -36,6 +36,7 @@ public final class TransactionSearchRequestHandlerTest {
         this.handler = new TransactionSearchRequestHandler(
                 supportComponent.viewController(),
                 factory.getContext().getAccountBook(),
+                supportComponent.accountBook().getRootAccount().getUUID(),
                 supportComponent.settings().getDateFormat(),
                 supportComponent.localizable());
     }